Moving vulkan forward declarations to a single file
